Organizando o mvc e adicionando o delete
