+++ WiredTwilightProject/WiredTwilightBackend/Model/UpdateUsernameRequest.cs
@@ -1,0 +1,7 @@
+namespace WiredTwilightBackend;
+
+using WiredTwilightBackend;
+public class UpdateUsernameRequest
+{
+    public string? NewUsername { get; set; }
+}
