--- WiredTwilightProject/WiredTwilightBackend/Tag.cs
@@ -1,15 +1,0 @@
-namespace WiredTwilightBackend;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-using BCrypt.Net;
-public class Tag
-{
-    [Key]
-    public int Id { get; set; }
-
-    [Required]
-    [MaxLength(50)]
-    public string Name { get; set; }  // Nome da tag
-
-    public ICollection<Post> Posts { get; set; }  // Posts associados à tag
-}
