+++ WiredTwilightProject/WiredTwilightBackend/appsettings.json
@@ -15,3 +15,4 @@
   },
   "AllowedHosts": "*"
 }
+
